Added Healthcare Avoidance Page and Codebook
This healthcare avoidance page was added as a project with the codebook and prebuilt html added within the hc_avoidance project. A tag of "hcavoid" was given without an eerg or covid tag to ensure it does not show up under either Covid-19 or Projects on the homepage. This makes it a sort of hidden page within the website.